Add uniform_size to constellations, fix random_color missing keys, float taper
- Constellation: add uniform_size option — all sub-objects in one instance
  share a single randomly-sampled size, works with r_min/max and radius_min/max
- random_color: default to black when color keys are absent (allows stroke_width=0
  profiles to omit stroke color keys without crashing)
- laserprep: change --taper argument type from int to float for subtle tapering
  on high layer-count pieces

+++ laserprep.py
@@ -365,7 +365,7 @@ def main():
     parser = argparse.ArgumentParser()
     parser.add_argument("--file", type=str, default=FILE_TO_PREP)
     parser.add_argument("--app", type=str, default=APPELLATION)
-    parser.add_argument("--taper", type=int, default=TAPER_PX)
+    parser.add_argument("--taper", type=float, default=TAPER_PX)
     args = parser.parse_args()
 
     canvas_size, shapes, base_name = load_shapes(args.file)
